refactor test files to remove dummy component tests and impmemnent forum question tests
